[kubevirt] Fix KubeVirtNodesLiveMigratable condition for NodePools with 0 replicas
In cases where the NodePool has 0 replicas and there are no KubevirtMachines registered with the nodepool, the condition of KubeVirtNodesLiveMigratable
should not be set at all (whether true or false).
This is done in order to fix the failure in the TestNodePool e2e, in which the HostedCluster conditions are being verified on a 0-replicas NodePool.

Signed-off-by: Oren Cohen <ocohen@redhat.com>